+++ src/main/kotlin/io/github/jamalam/database/DefaultValues.kt
@@ -89,6 +89,7 @@ fun KClass<ServerFilePasteConfig>.getDefault(): ServerFilePasteConfig {
 fun KClass<ServerPhishingConfig>.getDefault(): ServerPhishingConfig {
     return ServerPhishingConfig(
         enabled = false,
+        moderatorsExempt = true,
         moderationType = ServerPhishingModerationType.Delete,
     )
 }