+++ src/main/kotlin/io/github/jamalam/extensions/user/TagExtension.kt
@@ -17,6 +17,7 @@
 
 package io.github.jamalam.extensions.user
 
+import com.kotlindiscord.kord.extensions.checks.guildFor
 import com.kotlindiscord.kord.extensions.commands.Arguments
 import com.kotlindiscord.kord.extensions.commands.application.slash.ephemeralSubCommand
 import com.kotlindiscord.kord.extensions.commands.application.slash.publicSubCommand