+++ src/main/kotlin/io/github/jamalam/util/Constants.kt
@@ -17,6 +17,8 @@
 
 package io.github.jamalam.util
 
+import kotlinx.datetime.Clock
+
 /**
  * @author  Jamalam360
  */
@@ -25,5 +27,6 @@ const val PINGUINO_PFP = "https://images-ext-2.discordapp.net/external/tM2ezTNgh
         "/%3Fsize%3D256/https/cdn.discordapp.com/avatars/896758540784500797/507601ac" +
         "31f51ffc334fac125089f7ea.png"
 
-const val VERSION = "v0.7.3"
+const val VERSION = "v0.7.4"
 const val DBL_URL = "https://top.gg/api/bots/896758540784500797/stats"
+val BOOT_TIME = Clock.System.now()